# PHOENIX-6458 Using global indexes for queries with uncovered columns