--- a/phoenix-core/src/main/java/org/apache/phoenix/compile/ExpressionCompiler.java
+++ b/phoenix-core/src/main/java/org/apache/phoenix/compile/ExpressionCompiler.java
@@ -112,7 +112,7 @@ import org.apache.phoenix.schema.ColumnFamilyNotFoundException;
 import org.apache.phoenix.schema.ColumnNotFoundException;
 import org.apache.phoenix.schema.ColumnRef;
 import org.apache.phoenix.schema.DelegateDatum;
-import org.apache.phoenix.schema.LocalIndexDataColumnRef;
+import org.apache.phoenix.schema.IndexDataColumnRef;
 import org.apache.phoenix.schema.PColumn;
 import org.apache.phoenix.schema.PDatum;
 import org.apache.phoenix.schema.PTable;
@@ -139,6 +139,8 @@ import org.apache.phoenix.util.ExpressionUtil;
 import org.apache.phoenix.util.IndexUtil;
 import org.apache.phoenix.util.SchemaUtil;
 import org.apache.phoenix.util.StringUtil;
+
+import static org.apache.phoenix.util.IndexUtil.isHintedGlobalIndex;
 
 
 public class ExpressionCompiler extends UnsupportedAllParseNodeVisitor<Expression> {
@@ -371,9 +373,12 @@ public class ExpressionCompiler extends UnsupportedAllParseNodeVisitor<Expressio
             // Rather than not use a local index when a column not contained by it is referenced, we
             // join back to the data table in our coprocessor since this is a relatively cheap
             // operation given that we know the join is local.
-            if (context.getCurrentTable().getTable().getIndexType() == IndexType.LOCAL) {
+            if (context.getCurrentTable().getTable().getIndexType() == IndexType.LOCAL
+                    || isHintedGlobalIndex(context.getCurrentTable())) {
                 try {
-                    return new LocalIndexDataColumnRef(context, context.getCurrentTable(), node.getName());
+                    context.setUncoveredIndex(true);
+                    return new IndexDataColumnRef(context, context.getCurrentTable(),
+                            node.getName());
                 } catch (ColumnFamilyNotFoundException c) {
                     throw e;
                 }