--- a/phoenix-core/src/main/java/org/apache/phoenix/compile/FromCompiler.java
+++ b/phoenix-core/src/main/java/org/apache/phoenix/compile/FromCompiler.java
@@ -103,6 +103,7 @@ import org.apache.phoenix.thirdparty.com.google.common.collect.ListMultimap;
 import org.apache.phoenix.thirdparty.com.google.common.collect.Lists;
 
 import static org.apache.phoenix.monitoring.MetricType.NUM_METADATA_LOOKUP_FAILURES;
+import static org.apache.phoenix.util.IndexUtil.isHintedGlobalIndex;
 
 /**
  * Validates FROM clause and builds a ColumnResolver for resolving column references
@@ -1157,13 +1158,14 @@ public class FromCompiler {
     }
     
     private static class ProjectedTableColumnResolver extends MultiTableColumnResolver {
-        private final boolean isLocalIndex;
+        private final boolean isIndex;
         private final List<TableRef> theTableRefs;
         private final Map<ColumnRef, Integer> columnRefMap;
         private ProjectedTableColumnResolver(PTable projectedTable, PhoenixConnection conn, Map<String, UDFParseNode> udfParseNodes) throws SQLException {
             super(conn, 0, udfParseNodes, null);
             Preconditions.checkArgument(projectedTable.getType() == PTableType.PROJECTED);
-            this.isLocalIndex = projectedTable.getIndexType() == IndexType.LOCAL;
+            this.isIndex = projectedTable.getIndexType() == IndexType.LOCAL
+                    || projectedTable.getIndexType() == IndexType.GLOBAL;
             this.columnRefMap = new HashMap<ColumnRef, Integer>();
             long ts = Long.MAX_VALUE;
             for (int i = projectedTable.getBucketNum() == null ? 0 : 1; i < projectedTable.getColumns().size(); i++) {
@@ -1201,9 +1203,11 @@ public class FromCompiler {
             try {
                 colRef = super.resolveColumn(schemaName, tableName, colName);
             } catch (ColumnNotFoundException e) {
-                // This could be a ColumnRef for local index data column.
-                TableRef tableRef = isLocalIndex ? super.getTables().get(0) : super.resolveTable(schemaName, tableName);
-                if (tableRef.getTable().getIndexType() == IndexType.LOCAL) {
+                // This could be a ColumnRef for index data column.
+                TableRef tableRef = isIndex ? super.getTables().get(0)
+                        : super.resolveTable(schemaName, tableName);
+                if (tableRef.getTable().getIndexType() == IndexType.LOCAL
+                        || isHintedGlobalIndex(tableRef)) {
                     try {
                         TableRef parentTableRef = super.resolveTable(
                                 tableRef.getTable().getSchemaName().getString(),