… phoenix-core/src/main/java/org/apache/phoenix/compile/ProjectionCompiler.java
… phoenix-core/src/main/java/org/apache/phoenix/compile/TupleProjectionCompiler.java
… phoenix-core/src/main/java/org/apache/phoenix/coprocessor/BaseScannerRegionObserver.java
… phoenix-core/src/main/java/org/apache/phoenix/coprocessor/GroupedAggregateRegionObserver.java
… phoenix-core/src/main/java/org/apache/phoenix/execute/BaseQueryPlan.java
… phoenix-core/src/main/java/org/apache/phoenix/iterate/BaseResultIterators.java
… phoenix-core/src/main/java/org/apache/phoenix/iterate/ExplainTable.java
… phoenix-core/src/main/java/org/apache/phoenix/iterate/NonAggregateRegionScannerFactory.java
… phoenix-core/src/main/java/org/apache/phoenix/iterate/RegionScannerFactory.java
… phoenix-core/src/main/java/org/apache/phoenix/optimize/QueryOptimizer.java
… phoenix-core/src/main/java/org/apache/phoenix/schema/IndexDataColumnRef.java
… phoenix-core/src/main/java/org/apache/phoenix/schema/TableRef.java
… phoenix-core/src/main/java/org/apache/phoenix/util/IndexUtil.java
… phoenix-core/src/main/java/org/apache/phoenix/util/ScanUtil.java