@stoty 2021-07-25
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| +0 :ok: |  reexec  |   5m  0s |  Docker mode activated.  |
||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| +1 :green_heart: |  hbaseanti  |   0m  0s |  Patch does not have any anti-patterns.  |
| +1 :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| +1 :green_heart: |  test4tests  |   0m  0s |  The patch appears to include 2 new or modified test files.  |
||| _ master Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  22m 12s |  master passed  |
| +0 |  hbaserecompile  |  28m 37s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m  2s |  master passed  |
| +1 :green_heart: |  checkstyle  |   2m 26s |  master passed  |
| +1 :green_heart: |  javadoc  |   0m 51s |  master passed  |
| +0 :ok: |  spotbugs  |   3m  5s |  phoenix-core in master has 964 extant spotbugs warnings.  |
||| _ Patch Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  13m 29s |  the patch passed  |
| +0 |  hbaserecompile  |  23m 35s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m  6s |  the patch passed  |
| +1 :green_heart: |  javac  |   1m  6s |  the patch passed  |
| -1 :x: |  checkstyle  |   2m 36s |  phoenix-core: The patch generated 129 new + 4748 unchanged - 113 fixed = 4877 total (was 4861)  |
| +1 :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| +1 :green_heart: |  javadoc  |   0m 50s |  the patch passed  |
| -1 :x: |  spotbugs  |   3m 21s |  phoenix-core generated 1 new + 962 unchanged - 2 fixed = 963 total (was 964)  |
||| _ Other Tests _ |
| -1 :x: |  unit  | 115m 19s |  phoenix-core in the patch failed.  |
| +1 :green_heart: |  asflicense  |   0m 46s |  The patch does not generate ASF License warnings.  |
|  |   | 188m 42s |   |


| Reason | Tests |
|-------:|:------|
| FindBugs | module:phoenix-core |
|  |  org.apache.phoenix.schema.IndexDataColumnRef doesn't override ColumnRef.equals(Object)  At IndexDataColumnRef.java:At IndexDataColumnRef.java:[line 1] |
| Failed junit tests | phoenix.end2end.ViewIT |
|   | phoenix.end2end.DistinctPrefixFilterIT |
|   | phoenix.end2end.ExplainPlanWithStatsEnabledIT |
|   | phoenix.end2end.index.GlobalIndexOptimizationIT |
|   | phoenix.end2end.RowValueConstructorOffsetIT |
|   | phoenix.end2end.DefaultColumnValueIT |
|   | phoenix.end2end.index.IndexUsageIT |
|   | phoenix.end2end.IndexToolIT |
|   | phoenix.end2end.DerivedTableIT |
|   | phoenix.end2end.UserDefinedFunctionsIT |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.41 ServerAPI=1.41 base: https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/phoenix/pull/1256 |
| Optional Tests | dupname asflicense javac javadoc unit spotbugs hbaserebuild hbaseanti checkstyle compile |
| uname | Linux 34763af58f8a 4.15.0-112-generic #113-Ubuntu SMP Thu Jul 9 23:41:39 UTC 2020 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev/phoenix-personality.sh |
| git revision | master / 6842c1d |
| Default Java | Private Build-1.8.0_242-8u242-b08-0ubuntu3~16.04-b08 |
| checkstyle | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/artifact/yetus-general-check/output/diff-checkstyle-phoenix-core.txt |
| spotbugs | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/artifact/yetus-general-check/output/new-spotbugs-phoenix-core.html |
| unit | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/artifact/yetus-general-check/output/patch-unit-phoenix-core.txt |
|  Test Results | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/testReport/ |
| Max. process+thread count | 10804 (vs. ulimit of 30000) |
| modules | C: phoenix-core U: phoenix-core |
| Console output | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1256/4/console |
| versions | git=2.7.4 maven=3.3.9 spotbugs=4.1.3 |
| Powered by | Apache Yetus 0.12.0 https://yetus.apache.org |


This message was automatically generated.